+++ src/components/home/CoffeeSelection.tsx
@@ -1,25 +1,20 @@
-import { useResetRecoilState } from 'recoil';
-import { caffeineFilterState, registPostState } from '@/atoms/atoms';
-import CoffeeOptionSelection from '@/components/common/CoffeeOptionSelection';
-import CoffeeMenuSelection from '@/components/home/CoffeeMenuSelection';
 import Icon from '@/components/common/Icon';
+import CoffeeMenuSelection from '@/components/home/CoffeeMenuSelection';
+import CoffeeOptionSelection from '@/components/common/CoffeeOptionSelection';
 import { CAFFEINE_FILTER_TEXTS } from '@/constants/home';
 import { iconPropsGenerator } from '@/utils/iconPropsGenerator';
+import useResetSelectedCoffee from '@/hooks/useResetSelectedCoffee';
+
 import { cx } from 'styled-system/css';
 import { styled } from 'styled-system/jsx';
 import { Align, Between, Flex } from '@/styles/layout';
 import { SmStyle, SumBoardTitle, SumTitle } from '@/styles/styles';
+import React from 'react';
 
 const { title } = CAFFEINE_FILTER_TEXTS;
 
 const CoffeeSelection = () => {
-  const resetstate = useResetRecoilState(registPostState);
-  const resetCaffeineState = useResetRecoilState(caffeineFilterState);
-
-  const touchResetBtn = () => {
-    resetstate();
-    resetCaffeineState();
-  };
+  const touchResetBtn = useResetSelectedCoffee();
 
   return (
     <div className={SmStyle}>
